+++ src/components/App.js
@@ -9,7 +9,11 @@ import reducer from '../reducers/index';
 console.log({AppContext})
 
 const App =() => {
-  const [state, dispatch] = useReducer(reducer, [])
+  
+  const initialState = {
+    events: []
+  }
+  const [state, dispatch] = useReducer(reducer, initialState)
 
   return (
     <AppContext.Provider value={{state, dispatch}}>
