+++ src/components/EventForm.js
@@ -43,7 +43,7 @@ const EventForm = (props) => {
             </div>
 
             <button className="btn btn-primary" onClick={addEvent} disabled={unCreatable}>イベントを作成する</button>
-            <button className="btn btn-danger" onClick={deleteAllEvents} disabled={state.length === 0}>すべてのイベントを削除する</button>
+            <button className="btn btn-danger" onClick={deleteAllEvents} disabled={state.events.length === 0}>すべてのイベントを削除する</button>
             </form>
         </>
     )
